requirements.txt
easyocr==1.7.1
Pillow==10.0.1
Pillow==10.0.1
